python_flask_webscoket_sqlite_webcam/requirements.txt

pillow
opencv-python
flask==3.0.3
flask-socketio==4.3.2
python-engineio==3.13.2
Werkzeug==1.0.1
gevent
pandas
ultralytics
